pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_engineer_starter"
version = "0.0.1"
requires-python = ">=3.8"
dependencies = [
    "google-cloud-documentai==2.27.0",
    "google-generativeai==0.7.1",
    "pypdf==4.2.0",
    "python-dotenv==1.0.1",
]